Wire everything together to display a dynamic list using unidirectional data flow.
Flow the events up

Now that we have all the components we need – a ViewModel, a bridge composable TodoActivityScreen, and TodoScreen, we wired everything together to display a dynamic list using unidirectional data flow.

In TodoActivityScreen passed addItem and removeItem from the ViewModel.

Edited TodoActivityScreen to observe the todoItems LiveData using observeAsState