Icons shown and hidden automatically as the text changes state.
Run the app again and you'll see that the icons show and hide automatically as the text changes state. You can also change the icon selection. When you hit the "Add" button you will see that a new TodoItem is generated based on the values input.

Congratulations, you've learned about (state in compose, state hoisting, and) how to build dynamic UIs based on state.

In the next few sections we'll explore how to think about making reusable components that interact with state.